feat: use Listener trait
